move APM to core and fix issue #24

+++ Tools/BuildSrc/Package_Scripts/post_install.sh
@@ -14,15 +14,21 @@ fi
 
 ln -sf $(readlink -f ${ALUSUS_ROOT}/Bin/alusus) ${LINK_DIR}/alusus
 ln -sf $(readlink -f ${ALUSUS_ROOT}/Bin/الأسس) ${LINK_DIR}/الأسس
+ln -sf $(readlink -f ${ALUSUS_ROOT}/Bin/apm) ${LINK_DIR}/apm
 
 
 # Copy GTK syntax highlighting specs to gtksourceview folder.
 
+IS_GTKSOURCEVIEW_EXISTS=1
 LANG_SPEC_DIR="/usr/share/gtksourceview-3.0/language-specs"
 if [ -e "/usr/share/gnome/gtksourceview-3.0/language-specs" ]; then
   LANG_SPEC_DIR="/usr/share/gnome/gtksourceview-3.0/language-specs"
 elif [ -e "/usr/local/share/gtksourceview-3.0/language-specs" ]; then
   LANG_SPEC_DIR="/usr/local/share/gtksourceview-3.0/language-specs"
+else
+  IS_GTKSOURCEVIEW_EXISTS=0
 fi
 
-cp -f ${ALUSUS_ROOT}/Tools/Gtk_Syntax_Highlighting/alusus.lang ${LANG_SPEC_DIR}/
+if [ "$IS_GTKSOURCEVIEW_EXISTS" -eq 1 ]; then
+  cp -f ${ALUSUS_ROOT}/Tools/Gtk_Syntax_Highlighting/alusus.lang ${LANG_SPEC_DIR}/
+fi
